Properties mapped in our property mappers are overriden
Fixes #36927
