+++ quarkus/runtime/src/test/java/org/keycloak/quarkus/runtime/configuration/test/ConfigurationTest.java
@@ -531,35 +531,4 @@ public class ConfigurationTest extends AbstractConfigurationTest {
             assertEquals(Integer.toString(maxCount), config.getConfigValue(prop).getValue());
         }
     }
-
-    @Test
-    public void testWildcardCliOptionCanBeMappedToQuarkusOption() {
-        ConfigArgsConfigSource.setCliArgs("--log-level-org.keycloak=trace");
-        SmallRyeConfig config = createConfig();
-        assertEquals("TRACE", config.getConfigValue("quarkus.log.category.\"org.keycloak\".level").getValue());
-        assertEquals("INFO", config.getConfigValue("quarkus.log.category.\"io.quarkus\".level").getValue());
-        assertEquals("INFO", config.getConfigValue("quarkus.log.category.\"foo.bar\".level").getValue());
-    }
-
-    @Test
-    public void testWildcardEnvVarOptionCanBeMappedToQuarkusOption() {
-        putEnvVar("KC_LOG_LEVEL_IO_QUARKUS", "trace");
-        SmallRyeConfig config = createConfig();
-        assertEquals("INFO", config.getConfigValue("quarkus.log.category.\"org.keycloak\".level").getValue());
-        assertEquals("TRACE", config.getConfigValue("quarkus.log.category.\"io.quarkus\".level").getValue());
-        assertEquals("INFO", config.getConfigValue("quarkus.log.category.\"foo.bar\".level").getValue());
-    }
-
-    @Test
-    public void testWildcardOptionFromConfigFile() {
-        putEnvVar("SOME_CATEGORY_LOG_LEVEL", "debug");
-        SmallRyeConfig config = createConfig();
-        assertEquals("DEBUG", config.getConfigValue("quarkus.log.category.\"io.k8s\".level").getValue());
-    }
-
-    @Test
-    public void testWildcardPropertiesDontMatchEnvVarsFormat() {
-        SmallRyeConfig config = createConfig();
-        assertEquals("INFO", config.getConfigValue("quarkus.log.category.\"io.quarkus\".level").getValue());
-    }
 }
